log-fastapi-back/requirements.txt
fastapi
uvicorn[standard]
pymongo
PyJWT
python-decouple
python-jose==3.3.0
passlib==1.7.4
asyncpg==0.27.0
bcrypt==4.0.1
python-multipart==0.0.5